gfunction: now std layout and trivial copy
std::tuple cause problems when copied in to a SYCL lambda. The
strict spec requires standard layout and tirivally copyable,
which is not visible for all of gtensor and doesn't seem to be
actually necessary in practice. However std::tuple in
particular is problematic, so refactorying here to avoid it's
use.