add norm and conj for gt::complex
Requires some hackery to work with thrust. In particular, since CUDA
needs to use thrust::complex for ops to work even when thrust device
vector is not used as the storage backend, it also needs thrust_ext.h
for operations to behave, and somehow for the wrapping of norm and conf
to work.