blas: explicit dotu vs dotc
Overloading dot->dotu for complex could lead to confusion,
and is arguably inconsistent with the other mappings which
template only on the z/c/d/s prefix. This change follows
the convention of the MKL interface, and requires explicitly
specifying which form is desired for complex, dotu or dotc,
and dot is only for real numbers.