add "clib" namespace
This is to start a distinction between what's being used in gtensor itself,
vs what's used to generate the C/Fortran interface. The latter of which will
always be one compile-time defined space, while the rest might some day be
selectable via template. And there's already the option to use thrust for the
backend, in which case gtensor proper shouldn't have to worry about cuda vs hip
-- that's a different level underneath.