backend: separate gallocators by using, e.g., sycl, sycl_host, sycl_managed
If one really wants to run multiple backends, there's some iffiness in
that they do have slightly different ways of providing host memory,
though in the end they are all compatible, with the most significant
difference being that cuda will provide pinned host memory. In any case,
this gives the different labels and a more explicit choice as to what's
used for "host", anyway (by means of a type alias).