sycl: auto kernel name for launch, oneapi 2021-04
The 2021-04 release of oneapi won't compile with a lambda type
in the kernel name template. The only workaround I've found so
far is to use the new SYCL 2020 auto-generated kernel names. This
may make debugging and performance measurement more confusing,
so may be worth re-enabling later if the issue can be resolved.