test: improved test predicates
Adds overloaded GT_EXPECT_EQ(a, b), GT_EXPECT_NEAR(a, b),
and GT_EXPECT_NEAR_MAXERR(a, b, max_err) macros that work
with scalars, gtensor expressions, and combinations of the
two (expressions must be in first position). The idea is for
this to replace the confusingly named test_helpers.h header.