sparse: faster conversion, benchmark
Precompute row_ptr array, instead of just per batch nnz offsets.
This allows parallelizing on (row, batch), which is a huge
performance improvement, particularly for large single batch
conversion.

The exclusive scan to go from the per-row nnz counts (computed on
gpu) to the row_ptr offsets is done on host for now. Not sure
doing this on device will improve things much - still need to
copy the final element, nnz, to host, and nrows is likely going
to be small enough that it won't help. Doing a transform exclusive
scan so it's a single kernel probably would help, but is much
more complicated and requires using CUDA shared mem and other backend
equivalents to be efficient (or backend specific libs like CUB).