sparse: rename cuda backends, new only for cuda 12
Old API has better performance than generic until 12, and in 12
the old API was also removed. There is a bsrsm2 API which is very
similar and basically is csr when block size=1, which could be explored
for performance comparison, but the generic API seems more likely to
exist for a long time.