gview: add is_f_contiguous, data
The data method raises an exception or assert(0) in device code
if called on a non-forward-contiguous view. Note that this changes
semantics of has_container_methods to include view objects, and
allow them to be passed to gt-blas and gt-fft routines. It will
raise a runtime error if the view is not f contiguous. A compile
time error would be preferrable but it's not clear if this is possible
or practical.