solver: don't stage rhs and result in device buffers
saves space, and the only case where the staging seems to be benificial
AMD platforms when using managed memory. This can be more simply handled
with a wrapper class, which allocates buffers and wraps the solve of an
underlying solver implementation, so it only needs to be done once for
all solvers.